requirements.txt
numpy==1.21.6
pandas==1.3.5
Pillow==9.2.0
streamlit==1.12.0
streamlit_option_menu==0.3.2
xgboost==1.6.2
